+++ test/tstBoostGeometryAdapters.cpp
@@ -169,8 +169,8 @@ BOOST_AUTO_TEST_CASE(is_valid)
   BOOST_TEST(!bg::is_valid(other_invalid_box, message));
   BOOST_TEST(message == "Box has corners in wrong order");
 
-  auto const infty = std::numeric_limits<double>::infinity();
-  other_invalid_box = {{{1., 5., 3.}}, {{infty, 3., 6.}}};
+  auto const infty = std::numeric_limits<float>::infinity();
+  other_invalid_box = {{{1.f, 5.f, 3.f}}, {{infty, 3.f, 6.f}}};
   BOOST_TEST(!details::isValid(other_invalid_box));
   BOOST_TEST(!bg::is_valid(other_invalid_box, message));
   BOOST_TEST(message == "Geometry has point(s) with invalid coordinate(s)");
